+++ CHANGELOG.md
@@ -2903,3 +2903,4 @@
 - [2026-07-16 17:28] fix(perf): resolve race condition during perf initialization
 - [2026-07-16 17:49] feat(auth): implement responsive auth view component
 - [2026-07-16 19:57] feat(network): add bi-directional event bus handler for network
+- [2026-07-16 20:01] feat(cache): implement responsive cache view component